fix: Fixed bugs with package names and paths and go.mod updates, and fixed example YAML.
